pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-dispatch"
version = "0.1.0"
description = "Unified RL-based Dispatch and Rescheduling Policy for Autonomous Patrol Robots"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "YJP", email = "yjp@example.com"}
]
keywords = ["reinforcement-learning", "robotics", "patrol", "dispatch", "ppo", "navigation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.21.0",
    "torch>=2.0.0",
    "gymnasium>=0.28.0",
    "tensorboard>=2.13.0",
    "scipy>=1.7.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "dataclasses-json>=0.5.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]
ros = [
    "rclpy",
    "nav2-simple-commander",
    "geometry-msgs",
    "nav-msgs",
]

[project.urls]
Homepage = "https://github.com/yjp/rl_dispatch_mvp"
Repository = "https://github.com/yjp/rl_dispatch_mvp"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
